Refactored item codebase. Now new minecarts are a bit easier to add.
Sort of... To add a new Minecart Item, way easier. Just need to
implement the new minecart entity type, the entity itself, and then
create the `SoundTrackMinecartItem` with reference to the new minecart
type. Do what needs to be done to register the new entity and item, and
boom! Easy minecarts!